backend/requirements.txt
fastapi           # The API Framework
uvicorn           # The Server to run FastAPI
python-multipart  # For file uploading
pypdf             # To read PDF files
langchain         # To manage AI logic
langchain-openai  # To talk to GPT-4o
chromadb          # Vector Database (for RAG)
pydantic          # Data validation
python-dotenv     # To read .env files
sqlalchemy        # For the SQLite Database
psycopg2-binary  # <--- Add this line
asyncpg          # <--- Add this for async support (optional but good)
langchain-ollama
tiktoken
langchain-text-splitters